requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
distlib==0.3.6
filelock==3.9.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
pipenv==2023.2.18
platformdirs==3.1.1
protobuf==3.20.3
py-mon==1.1.0
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-telegram-bot==20.1
python-telegram-bot-calendar==1.0.5
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
telebot==0.0.5
telegram==0.0.1
urllib3==1.26.15
virtualenv==20.21.0
virtualenv-clone==0.5.7
watchdog==2.3.1
